python: better site-packages and install-scripts
Previous versions of the Python 2.x formula did not set up the
site-packages and install-scripts folders in a way friendly to updates
of Python itself.

Though Python 2.7.1 *may* be the last 2.7.x release, these paths are
being redone in the case that 2.7.2 comes out some day, and to allow
the removal of the pip and distribute formulae.